ensure rhmi cr is created first before the operator deployment when testing olm e2e
When running the target test/e2e/olm, the operator deployed creates the default rhmi cr before the target deploy/integreatly-rhmi-cr.yaml is called. If an installation prefix is specified when running this target, the installation will fail as the smtp secret would have been created with a different prefix than what is specified in the rhmi cr by default.

This change ensures that the target deploy/integreatly-rhmi-cr.yaml is called before the operator can create the default rhmi cr.